backend/doc_extractor/requirements.txt
# Core Web Framework
fastapi>=0.119.0
uvicorn>=0.38.0
python-multipart>=0.0.20

# PDF Processing
pypdf>=6.0.0
pdf2image>=1.16.0
pdfplumber>=0.9.0
pytesseract>=0.3.10

# OCR & Computer Vision
paddleocr>=3.3.0
paddlepaddle>=3.0.0
opencv-contrib-python>=4.5.0
pillow>=9.0.0
paddlex[ocr]>=3.3.0

# Language Model
langchain>=0.2.14
langchain-core>=0.2.43
langchain-text-splitters>=0.2.2
langchain-google-genai>=1.0.5
langchain-ollama>=0.1.3
google-generativeai>=0.5.4

# Utilities
python-dotenv>=1.1.0
pydantic>=2.12.0
numpy>=1.26.0
